fix: Improve directory deletion on Windows
- Implement custom rmtree function to handle Windows deletion errors
- Replace shutil.rmtree calls with new rmtree utility function
- Move __onerror handler to utils.py for centralized error handling
